requirements.txt
pbr>=1.6
anyjson>=0.3.3
eventlet>=0.17.4
PyMySQL>=0.6.2  # MIT License
Pillow==2.4.0 # MIT
requests>=2.5.2
Flask-Bootstrap
Flask>=0.10,<1.0
flask-restless
flask-sqlalchemy
oslo.config>=2.3.0  # Apache-2.0
oslo.log>=1.8.0  # Apache-2.0
PrettyTable>=0.7,<0.8
python-swiftclient
python-keystoneclient
kombu>=3.0.7
